fix training_step in base model in mini test

+++ application/jobs/minimal_training_pytorch_cnn/app/custom/models/base_model.py
@@ -31,7 +31,7 @@ class VeryBasicModel(pl.LightningModule):
         x = batch["source"]
         y = batch["target"]
 
-        y = y.unsqueeze(1)
+        y = y.unsqueeze(1).float()
 
         y_hat = self(x)
         loss = self.loss(y_hat, y)
